task1/requirements.txt
colorama==0.4.6
Faker==37.4.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
tzdata==2025.2
